gouvdown HTML document

gouvdown HTML document

html_gouv(
  extra_dependencies = list(),
  includes = NULL,
  ...,
  logo = NULL,
  use_gouvdown_fonts = TRUE
)

Arguments

extra_dependencies
Additional function arguments to pass to the base R Markdown HTML output formatter html_document_base

includes
Named list of additional content to include within the document (typically created using the includes function).

...
Additional arguments passed to rmarkdown::html_document().

logo
a logo available in gouvdown::list_logos() or a file path to image file

use_gouvdown_fonts
Use the fonts provided by the gouvdown.fonts package, if installed.

Value

An R Markdown output format object to be passed to rmarkdown::render().